requirements.txt
streamlit
Pillow
requests
beautifulsoup4
lxml
fake_useragent
numpy
torch
torchvision
facenet-pytorch
wikipedia-api
lxml
fpdf2
